refactored RestApi
