+++ registry/RestApi/src/index.js
@@ -1,124 +1,43 @@
-import { join as joinPath } from 'path'
-import { forEach, resolve, reduce } from '@serverless/utils'
-
-const catchallParameterPattern = /{\.{3}([^}]+?)}/g
-
-async function getAwsApiGatewayInputs(inputs) {
-  const apiGatewayInputs = {
-    apiName: inputs.apiName,
-    role: inputs.role,
-    routes: {}
-  }
-
-  apiGatewayInputs.routes = reduce(
-    (pathAcc, methods, path) => {
-      const reparameterizedPath = path.replace(catchallParameterPattern, '{$1+}')
-      const normalizedPath = reparameterizedPath.replace(/^\/+/, '')
-      const methodDefinitions = reduce(
-        (methodAcc, methodObject, method) => {
-          const normalizedMethod = method.toUpperCase()
-          methodAcc[normalizedMethod] = methodObject
-          return methodAcc
-        },
-        {},
-        methods
-      )
-      pathAcc[normalizedPath] = methodDefinitions
-      return pathAcc
-    },
-    {},
-    inputs.routes
-  )
-
-  return apiGatewayInputs
-}
-
-async function constructApiGateway(inputs, context, provider) {
-  const apiInputs = await getAwsApiGatewayInputs(inputs)
-
-  const apiGatewayComponent = await context.import('AwsApiGateway')
-  const apiGateway = context.construct(apiGatewayComponent, {
-    ...apiInputs,
-    provider: provider
-  })
-  return apiGateway
-}
-
-function flattenRoutes(routes) {
-  const flattened = {}
-  function doFlatten(subRoutes, basePath) {
-    forEach((valueO, keyO) => {
-      const key = resolve(keyO)
-      const value = resolve(valueO)
-      if (key.startsWith('/')) {
-        doFlatten(value, joinPath(basePath, key))
-      } else {
-        if (key.toLowerCase() === 'routes') {
-          return doFlatten(value, basePath)
-        } else if (
-          !['any', 'delete', 'get', 'head', 'options', 'patch', 'post', 'put'].includes(
-            key.toLowerCase()
-          )
-        ) {
-          throw new Error(
-            `Configuration key "${key}" was interpreted as an HTTP method because it does not start with a slash, but it is not a valid method.`
-          )
-        }
-        if (flattened[basePath]) {
-          if (flattened[basePath][key]) {
-            throw new Error(`Method "${key}" on route [${basePath}] was mapped more than once.`)
-          }
-        } else {
-          flattened[basePath] = {}
-        }
-        flattened[basePath][key] = value
-      }
-    }, subRoutes)
-  }
-  doFlatten(routes, '/')
-
-  return flattened
-}
+import { resolve } from '@serverless/utils'
+import { flattenRoutes, normalizeRoutes, getSwaggerDefinition } from './utils'
+import { newVariable } from '../../../dist/utils/variable'
 
 const RestApi = async function(SuperClass, SuperContext) {
   const iamComponent = await SuperContext.import('AwsIamRole')
+  const AwsApiGatewayComponent = await SuperContext.import('AwsApiGateway')
 
   return class extends SuperClass {
     async define(context) {
-      const { inputs } = this
-      if (!['AwsApiGateway'].includes(resolve(inputs.gateway))) {
+      const provider = resolve(this.provider)
+      if (!['AwsApiGateway'].includes(resolve(this.gateway))) {
         throw new Error('Specified "gateway" is not supported.')
       }
 
-      const provider = resolve(inputs.provider)
-      const flatRoutes = flattenRoutes(resolve(inputs.routes))
-      const childComponents = []
-      const name = `${resolve(inputs.apiName)}-iam-role`
+      const roleName = `${resolve(this.apiName)}-iam-role`
       const service = 'apigateway.amazonaws.com'
-      this.role = context.construct(iamComponent, {
-        roleName: name,
-        service,
-        provider
-      })
-      childComponents.push(this.role)
 
-      const flatInputs = { ...inputs, routes: flatRoutes }
-      this.gateway = await constructApiGateway(
-        {
-          ...flatInputs,
-          role: this.role // TODO: add functionality to read from state so that update works
-        },
-        context,
-        provider
-      )
-      childComponents.push(this.gateway)
+      // eslint does not allow destructuring because this is a reserved word...
+      const roleArn = newVariable('${this.children.role.arn}', { this: this })
+      const flatRoutes = flattenRoutes(resolve(this.routes))
+      const normalizedRoutes = normalizeRoutes(flatRoutes)
+      const swaggerTemplate = getSwaggerDefinition(resolve(this.apiName), roleArn, normalizedRoutes)
 
-      return childComponents
+      return {
+        role: context.construct(iamComponent, {
+          roleName,
+          service,
+          provider
+        }),
+        gateway: context.construct(AwsApiGatewayComponent, {
+          provider,
+          swaggerTemplate
+        })
+      }
     }
 
     async deploy() {
       Object.assign(this, {
-        apiName: resolve(this.inputs.apiName),
+        apiName: resolve(this.apiName),
         paths: this.gateway.urls,
         baseUrl: this.gateway.baseUrl,
         gateway: {
